+++ .fitnessrc.ts
@@ -1,8 +1,9 @@
 export default {
   checks: [
-    'changelog',
     'changelog-updated',
+    'changelog',
     'cspell',
+    'markdown-front-matter',
     'node-version',
     'semantic-commit',
   ],